Extract card component and use it in Learn and Event pages (#961)
