+++ components/events/EventCard.vue
@@ -1,54 +1,32 @@
 <template>
-  <article class="event-card">
-    <div
-      class="event-card__image"
-      :lazy-background="image"
-    />
-    <div class="event-card__content">
-      <header class="event-card__header">
-        <h3 class="event-card__title">
-          {{ title }}
-        </h3>
-        <div class="event-card__tags">
-          <cv-tag
-            v-for="type in types"
-            :key="type"
-            :label="type"
-            kind="purple"
-          />
-        </div>
-      </header>
-
-      <div class="event-card__info-detail">
-        <p class="event-card__location">
-          <Map20 class="event-card__icon" />
-          {{ location }}
-        </p>
-        <p class="event-card__date">
-          <Calendar20 class="event-card__icon" />
-          <time>{{ date }}</time>
-        </p>
-      </div>
-
-      <footer class="event-card__link">
-        <AppCta
-          v-if="hasWebsite"
-          :url="to"
-        >
-          Join the event
-        </AppCta>
-      </footer>
+  <AppCard
+    class="event-card"
+    :image="image"
+    :title="title"
+    :tags="types"
+    :to="to"
+    cta-label="Join the event"
+  >
+    <div>
+      <p class="event-card__location">
+        <Map20 class="event-card__icon" />
+        {{ location }}
+      </p>
+      <p class="event-card__date">
+        <Calendar20 class="event-card__icon" />
+        <time>{{ date }}</time>
+      </p>
     </div>
-  </article>
+  </AppCard>
 </template>
 
 <script lang="ts">
 import Vue from 'vue'
 import { Component, Prop } from 'vue-property-decorator'
-import AppCta from '~/components/ui/AppCta.vue'
+import AppCard from '~/components/ui/AppCard.vue'
 
 @Component({
-  components: { AppCta }
+  components: { AppCard }
 })
 export default class extends Vue {
   @Prop(Array) types!: any
@@ -57,114 +35,28 @@ export default class extends Vue {
   @Prop(String) location!: any
   @Prop(String) date!: any
   @Prop(String) to!: any
-
-  hasWebsite: boolean = !!this.to
 }
 </script>
 
 <style lang="scss" scoped>
 @import '~carbon-components/scss/globals/scss/typography';
 
-$img-medium-width: 13rem;
-
 .event-card {
-  min-height: 13rem;
-  width: 100%;
   margin-bottom: $layout-03;
-  background-color: $cool-gray-10;
-  color: $cool-gray-80;
-  display: flex;
-
-  @include mq($until: medium) {
-    height: auto;
-    flex-direction: column;
-  }
-
-  &__image {
-    flex: 0 0 14rem;
-    background-color: $cool-gray-80;
-    background-repeat: no-repeat;
-    background-size: 100%;
-    background-position: center;
-    overflow: hidden;
-
-    @include mq($from: medium, $until: large) {
-      flex: 0 0 13rem;
-    }
-
-    @include mq($until: medium) {
-      height: 13rem;
-      width: auto;
-    }
-  }
-
-  &__content {
-    @include type-style('body-long-01');
-    padding: $spacing-05 $spacing-05 $spacing-05 $spacing-07;
-  }
-
-  &__header {
-    display: flex;
-    justify-content: space-between;
-    align-items: flex-start;
-
-    @include mq($until: large) {
-      flex-direction: column;
-    }
-  }
-
-  &__title {
-    @include type-style('productive-heading-02');
-    flex: 0 0 auto;
-  }
-
-  &__tags {
-    width: 20rem;
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: flex-end;
-
-    @include mq($until: large) {
-      justify-content: flex-start;
-      margin-top: $spacing-03;
-    }
-
-    @include mq($from: medium, $until: large) {
-      width: auto;
-    }
-
-    @include mq($until: medium) {
-      width: 100%;
-    }
-
-    .bx--tag--purple {
-      background-color: $purple-70;
-      color: $white;
-    }
-
-    .bx--tag:first-child {
-      margin-left: 0;
-    }
-  }
-
-  &__info-detail {
-    margin-top: $layout-02;
-  }
 
   &__location, &__date {
     @include type-style('body-long-01');
-    margin-bottom: $spacing-03;
     display: flex;
     align-items: center;
+  }
+
+  &__location {
+    margin-bottom: $spacing-03;
   }
 
   &__icon {
     margin-right: $spacing-03;
     fill: currentColor;
-  }
-
-  &__link {
-    margin-top: $layout-03;
   }
 }
 </style>
